# Optimize historical NFT count performance